refactor: button.tsx to js
